Added Product Ready Alerts / Add Photo functionality
Alerts now display at the top of the page whenever a product is being searched. When the search is ready, the alert changes to "Success", and can be clicked on to display the "select product" modals.

In addition, the add photo to listing functionality has been completed. Photos are uploaded to the server and stored in a temp folder to be displayed in the form.

Tons of node_module files added